backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
Faker==37.12.0
fastapi==0.121.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.9.1
prisma==0.15.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
